+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-24T14:55:59.713Z",
+    "2026-03-25T05:23:50.764Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {