+++ packages/orq-rc/RELEASES.md
@@ -7739,3 +7739,13 @@ Based on:
 - [typescript v4.7.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.17 - packages/orq-rc
+
+## 2026-03-25 05:26:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.2 (2.869.23) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.18 - packages/orq-rc